backend/requirements.txt
flask
flask-bcrypt
flask-cors
pymongo[srv]
python-dotenv
pyjwt
anthropic
openai
requests
dnspython
twilio
edge_tts
gunicorn
numpy
opencv-python
face_recognition
SpeechRecognition
pydub
